cambios en admin

+++ controllers/productController.js
@@ -1,60 +1,60 @@
 const Product = require('../models/Product');
+const mongoose = require('mongoose');
 
-// Obtener todos los productos
 exports.getProducts = async (req, res) => {
   try {
     const products = await Product.find();
     res.json(products);
   } catch (error) {
-    console.error('Error al obtener productos:', error);
-    res.status(500).json({ message: 'Error al obtener productos', error: error.message });
+    res.status(500).json({ message: error.message });
   }
 };
 
-// Obtener producto por ID
 exports.getProduct = async (req, res) => {
   try {
+    if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+        return res.status(404).json({ message: 'ID inválido' });
+    }
     const product = await Product.findById(req.params.id);
     if (!product) return res.status(404).json({ message: 'Producto no encontrado' });
     res.json(product);
   } catch (error) {
-    console.error('Error al obtener producto:', error);
-    res.status(500).json({ message: 'Error al obtener producto', error: error.message });
+    res.status(500).json({ message: error.message });
   }
 };
 
-// Crear producto
 exports.createProduct = async (req, res) => {
   try {
     const newProduct = new Product(req.body);
     await newProduct.save();
     res.status(201).json(newProduct);
   } catch (error) {
-    console.error('Error al crear producto:', error);
     res.status(500).json({ message: 'Error al crear producto', error: error.message });
   }
 };
 
-// Actualizar producto
 exports.updateProduct = async (req, res) => {
   try {
+    if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+        return res.status(400).json({ message: 'ID inválido' });
+    }
     const updatedProduct = await Product.findByIdAndUpdate(req.params.id, req.body, { new: true });
     if (!updatedProduct) return res.status(404).json({ message: 'Producto no encontrado' });
     res.json(updatedProduct);
   } catch (error) {
-    console.error('Error al actualizar producto:', error);
-    res.status(500).json({ message: 'Error al actualizar producto', error: error.message });
+    res.status(500).json({ message: error.message });
   }
 };
 
-// Eliminar producto
 exports.deleteProduct = async (req, res) => {
   try {
+    if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+        return res.status(400).json({ message: 'ID inválido' });
+    }
     const deletedProduct = await Product.findByIdAndDelete(req.params.id);
     if (!deletedProduct) return res.status(404).json({ message: 'Producto no encontrado' });
     res.json({ message: 'Producto eliminado' });
   } catch (error) {
-    console.error('Error al eliminar producto:', error);
-    res.status(500).json({ message: 'Error al eliminar producto', error: error.message });
+    res.status(500).json({ message: error.message });
   }
 };
